proof of concept done. Tested with Y Axis motor

+++ Marlin/stepper.h
@@ -94,11 +94,11 @@ extern block_t *current_block;  // A pointer to the block currently being traced
 
 void quickStop();
 
-void digitalPotWrite(int address, int value);
+void digitalPotWrite(uint8_t address, uint8_t value);
 void microstep_ms(uint8_t driver, int8_t ms1, int8_t ms2);
 void microstep_mode(uint8_t driver, uint8_t stepping);
 void digipot_init();
-void digipot_current(uint8_t driver, int current);
+void digipot_current(uint8_t driver, uint8_t current);
 void microstep_init();
 void microstep_readings();
